need to check first if its there before can check if empty

+++ src/mjolnir/pbfgraphbuilder.cc
@@ -108,7 +108,7 @@ int main(int argc, char** argv) {
   valhalla::baldr::TileHierarchy hierarchy(tile_dir);
   for(const auto& level : hierarchy.levels()) {
     auto level_dir = tile_dir + "/" + std::to_string(level.first);
-    if(!boost::filesystem::is_empty(level_dir)) {
+    if(boost::filesystem::exists(level_dir) && !boost::filesystem::is_empty(level_dir)) {
       LOG_WARN("Non-empty " + level_dir + " will be purged of tiles");
       boost::filesystem::remove_all(level_dir);
     }
